requirements.txt
mss==9.0.1
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
llama-cpp-python==0.2.56
python-dotenv==1.0.0
pyyaml==6.0.1